2019 08 06
Player이동 추가
checkPoint 추가
가속도 추가
GameManager 추가
포물선 실패

+++ Limbo/Limbo/EndScene.cpp
@@ -21,7 +21,7 @@ void EndScene::Control()
 
 }
 
-void EndScene::Update(int)
+void EndScene::Update(float)
 {
 
 }
