+++ Limbo/Limbo/EndScene.h
@@ -10,6 +10,6 @@ public:
 
 	void Init();
 	void Control();
-	void Update(int);
+	void Update(float);
 	void Render(Gdiplus::Graphics*);
 };
