+++ Limbo/Limbo/GameScene.h
@@ -11,7 +11,7 @@ public:
 	
 	void Init();
 	void Control();
-	void Update(int);
+	void Update(float);
 	void Render(Gdiplus::Graphics*);
 
 private:
