+++ Limbo/Limbo/Scene.h
@@ -8,7 +8,7 @@ public:
 
 	virtual void Init() = 0;
 	virtual void Control() = 0;
-	virtual void Update(int) = 0;
+	virtual void Update(float) = 0;
 	virtual void Render(Gdiplus::Graphics*) = 0;
 
 };
